Relocate all errors related to imports
