+++ src/modules/errors.ts
@@ -13,6 +13,7 @@ import { UNKNOWN_LOCATION } from '../constants'
 
 import { RuntimeSourceError } from '../errors/runtimeSourceError'
 import { ErrorSeverity, ErrorType, SourceError } from '../types'
+import { nonAlphanumericCharEncoding } from './preprocessor/filePaths'
 
 export class ModuleConnectionError extends RuntimeSourceError {
   private static message: string = `Unable to get modules.`
@@ -105,6 +106,7 @@ export class UndefinedDefaultImportError extends UndefinedImportErrorBase {
     return `'${this.moduleName}' does not contain a default export!`
   }
 }
+
 export class UndefinedNamespaceImportError extends UndefinedImportErrorBase {
   constructor(moduleName: string, node?: ImportNamespaceSpecifier | ExportAllDeclaration) {
     super(moduleName, node)
@@ -164,5 +166,40 @@ export class CircularImportError implements SourceError {
 
   public elaborate() {
     return 'Break the circular import cycle by removing imports from any of the offending files.'
+  }
+}
+
+export abstract class InvalidFilePathError implements SourceError {
+  public type = ErrorType.TYPE
+  public severity = ErrorSeverity.ERROR
+  public location = UNKNOWN_LOCATION
+
+  constructor(public filePath: string) {}
+
+  abstract explain(): string
+
+  abstract elaborate(): string
+}
+
+export class IllegalCharInFilePathError extends InvalidFilePathError {
+  public explain() {
+    const validNonAlphanumericChars = Object.keys(nonAlphanumericCharEncoding)
+      .map(char => `'${char}'`)
+      .join(', ')
+    return `File path '${this.filePath}' must only contain alphanumeric chars and/or ${validNonAlphanumericChars}.`
+  }
+
+  public elaborate() {
+    return 'Rename the offending file path to only use valid chars.'
+  }
+}
+
+export class ConsecutiveSlashesInFilePathError extends InvalidFilePathError {
+  public explain() {
+    return `File path '${this.filePath}' cannot contain consecutive slashes '//'.`
+  }
+
+  public elaborate() {
+    return 'Remove consecutive slashes from the offending file path.'
   }
 }
